+++ Implementierung/moore.h
@@ -1,5 +1,5 @@
-void moore_c_iterative(uint64_t degree, uint64_t *x, uint64_t *y);
-void moore_c_batch(uint64_t degree, uint64_t *x, uint64_t *y);
+void moore_c_iterative(uint32_t degree, uint32_t *x, uint32_t *y);
+void moore_c_batch(uint32_t degree, uint32_t *x, uint32_t *y);
 
 
-extern void moore(uint64_t degree, uint64_t *x, uint64_t *y);
+extern void moore(uint32_t degree, uint32_t *x, uint32_t *y);
